requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
black==23.3.0
bs4==0.0.1
cachelib==0.9.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
construct==2.0.6
cryptography==40.0.2
cssselect==1.2.0
dash==2.9.3
dash-core-components==2.0.0
dash-extensions==1.0.0
dash-html-components==2.0.0
dash-iconify==0.1.2
dash-labs==1.2.0
dash-mantine-components==0.12.1
dash-table==5.0.0
dnspython==2.3.0
dominate==2.7.0
EditorConfig==0.12.3
email-validator==2.0.0.post2
et-xmlfile==1.1.0
fake-useragent==1.1.3
Flask==2.2.2
Flask-Admin==1.6.1
Flask-Bootstrap==3.3.7.1
Flask-Caching==2.0.2
Flask-Login==0.6.2
Flask-Redmail==0.3.0
Flask-Session==0.4.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
frozendict==2.3.8
greenlet==2.0.2
gunicorn==20.1.0
html5lib==1.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsbeautifier==1.14.7
kurt==2.0.7
lxml==4.9.2
MarkupSafe==2.1.2
more-itertools==9.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
parse==1.19.0
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.0
plotly==5.14.1
pycparser==2.21
pyee==8.2.2
PyJWT==2.6.0
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redmail==0.6.0
requests==2.30.0
requests-html==0.10.0
simplejson==3.19.1
six==1.16.0
skip==0.1.0
soupsieve==2.4.1
SQLAlchemy==2.0.12
tenacity==8.2.2
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
visitor==0.1.3
w3lib==2.1.1
webencodings==0.5.1
websockets==10.4
Werkzeug==2.3.4
WTForms==3.0.1
yahoo-finance==1.4.0
yfinance==0.2.18
zipp==3.15.0
